lesson8
* Добавлены фоновая музыка и звук монетки.
* Добавлены экраны.
