+++ assets/maps/map2 — копия.tmx
@@ -1,5 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <map version="1.8" tiledversion="1.8.5" orientation="orthogonal" renderorder="right-down" width="165" height="108" tilewidth="16" tileheight="16" infinite="0" nextlayerid="14" nextobjectid="60">
+ <properties>
+  <property name="fon" value="back.jpg"/>
+ </properties>
  <tileset firstgid="1" name="rLJjmOe" tilewidth="16" tileheight="16" tilecount="1680" columns="56">
   <image source="rLJjmOe.png" width="896" height="480"/>
  </tileset>
